Added code that I hope will connect the arduino to wifi
Added code that I hope will connect the arduino to wifi. This only works for unencrypted networks and you'll have to modify the SSID of your network. Also, this library assumes that you're using a shield and you have to use 10, 11, 12, and 13 for the SPI connection to the HDG104 module. I don't have a wifi capable arduino yet, so I'll probably need to make changes once that comes in.