Added a call to get time from an NTP server.
Added a call to get time from an NTP server. 

We just want to use this to try to schedule the pump to come on during the night to use electricity during off-peak hours.

There's still no code in here to actually do anything with the time, waiting on the wifi enabled arduino. After I get the new board, I'll start plunking around with figuring out what to do with the time.

Most of the added code came from https://www.arduino.cc/en/Tutorial/UdpNTPClient